separated parameters tests to own file
